+++ tests/fixtures/sleeper.py
@@ -4,29 +4,39 @@ This script can be used to test signal handling.
 """
 
 import logging
+import sys
 import time
 
-from yapw import clients
-from yapw.methods.blocking import ack
+import pika
+
+from yapw.clients import AsyncConsumer, Blocking
+from yapw.methods import ack
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
 
 
-class Client(clients.Blocking):
-    pass
-
-
 def callback(state, channel, method, properties, body):
     logger.info("Sleep")
-    time.sleep(10)
+    time.sleep(5)
     logger.info("Wake!")
-    ack(state, channel, method.delivery_tag)
+    if isinstance(state.connection, pika.BlockingConnection):
+        ack(state, channel, method.delivery_tag)
+    else:
+        channel.basic_ack(method.delivery_tag)
 
 
 def main():
-    client = Client(durable=False, exchange="yapw_development")
-    client.consume(callback, "sleep")
+    kwargs = {"durable": False, "exchange": "yapw_development", "prefetch_count": 10}
+    consumer_kwargs = {"callback": callback, "queue": "sleep"}
+    if len(sys.argv) > 1 and sys.argv[1] == "Blocking":
+        print("Blocking consumer")
+        client = Blocking(**kwargs)
+        client.consume(**consumer_kwargs)
+    else:
+        print("Asynchronous consumer")
+        client = AsyncConsumer(**kwargs, **consumer_kwargs)
+        client.start()
 
 
 if __name__ == "__main__":